+++ applications/fishing-map/src/features/map/Map.tsx
@@ -5,7 +5,7 @@ import { MapLegend } from '@globalfishingwatch/ui-components/dist'
 import type { Map } from '@globalfishingwatch/mapbox-gl'
 import { InteractiveMap, MapRequest } from '@globalfishingwatch/react-map-gl'
 import GFWAPI from '@globalfishingwatch/api-client'
-import useTilesState from '@globalfishingwatch/react-hooks/dist/use-tiles-state'
+import useTilesLoading from '@globalfishingwatch/react-hooks/dist/use-tiles-loading'
 import useLayerComposer from '@globalfishingwatch/react-hooks/dist/use-layer-composer'
 import {
   useMapClick,
@@ -58,7 +58,10 @@ const transformRequest: (...args: any[]) => MapRequest = (url: string, resourceT
 }
 
 const handleError = ({ error }: any) => {
-  if (error?.status === 401 && error?.url.includes('globalfishingwatch')) {
+  if (
+    (error?.status === 401 || error?.status === 403) &&
+    error?.url.includes('globalfishingwatch')
+  ) {
     GFWAPI.refreshAPIToken()
   }
 }
@@ -185,7 +188,7 @@ const MapWrapper = (): React.ReactElement | null => {
   const resetHoverState = useCallback(() => {
     setHoveredEvent(null)
     setHoveredDebouncedEvent(null)
-    cleanFeatureState()
+    cleanFeatureState('hover')
   }, [cleanFeatureState])
 
   const { viewport, onViewportChange } = useViewport()
@@ -220,7 +223,7 @@ const MapWrapper = (): React.ReactElement | null => {
 
   // TODO handle also in case of error
   // https://docs.mapbox.com/mapbox-gl-js/api/map/#map.event:sourcedataloading
-  const tilesLoading = useTilesState(mapInstance as Map)
+  const tilesLoading = useTilesLoading(mapInstance as Map)
 
   useEffect(() => {
     if (mapInstance) {
@@ -279,7 +282,7 @@ const MapWrapper = (): React.ReactElement | null => {
           <MapInfo center={hoveredEvent} />
         </InteractiveMap>
       )}
-      <MapControls onMouseEnter={resetHoverState} mapLoading={tilesLoading.loading} />
+      <MapControls onMouseEnter={resetHoverState} mapLoading={tilesLoading} />
       {layersWithLegend?.map((legend) => {
         const legendDomElement = document.getElementById(legend.id as string)
         if (legendDomElement) {