+++ applications/fishing-map/src/features/map/map.hooks.ts
@@ -1,8 +1,14 @@
 import { useSelector, useDispatch } from 'react-redux'
 import { useRef } from 'react'
-import { ExtendedFeatureVessel, InteractionEvent } from '@globalfishingwatch/react-hooks'
+import { Geometry } from 'geojson'
+import {
+  ExtendedFeatureVessel,
+  InteractionEvent,
+  useTilesLoading,
+} from '@globalfishingwatch/react-hooks'
 import { Generators, TimeChunks } from '@globalfishingwatch/layer-composer'
 import { ContextLayerType, Type } from '@globalfishingwatch/layer-composer/dist/generators/types'
+import { Style } from '@globalfishingwatch/mapbox-gl'
 import {
   selectDataviewInstancesResolved,
   selectTemporalgridDataviews,
@@ -47,6 +53,7 @@ export const useClickedEventConnect = () => {
   const rulersEditing = useSelector(selectEditing)
 
   const dispatchClickedEvent = (event: InteractionEvent | null) => {
+    // Used on workspaces-list or user panel to go to the workspace detail page
     if (locationType === USER || locationType === WORKSPACES_LIST) {
       const workspace = event?.features?.find(
         (feature: any) => feature.properties.type === 'workspace'
@@ -107,8 +114,11 @@ export type TooltipEventFeature = {
   type?: Type
   color?: string
   unit?: string
+  source: string
+  sourceLayer: string
   layerId: string
   contextLayer?: ContextLayerType | null
+  geometry?: Geometry
   value: string
   properties: Record<string, string>
   vesselsInfo?: {
@@ -151,6 +161,8 @@ export const useMapTooltip = (event?: InteractionEvent | null) => {
       // Not needed to create a dataview just for the workspaces list interaction
       if (feature.generatorId && (feature.generatorId as string).includes(WORKSPACE_GENERATOR_ID)) {
         const tooltipWorkspaceFeature: TooltipEventFeature = {
+          source: feature.source,
+          sourceLayer: feature.sourceLayer,
           layerId: feature.layerId as string,
           type: Generators.Type.GL,
           value: feature.properties.label,
@@ -167,6 +179,9 @@ export const useMapTooltip = (event?: InteractionEvent | null) => {
       color: dataview.config?.color || 'black',
       unit: feature.unit,
       value: feature.value,
+      source: feature.source,
+      sourceLayer: feature.sourceLayer,
+      geometry: feature.geometry,
       layerId: feature.layerId,
       contextLayer: feature.generatorContextLayer,
       properties: { ...feature.properties },
@@ -206,9 +221,13 @@ export const useMapTooltip = (event?: InteractionEvent | null) => {
 
 export const useMapStyle = () => {
   const mapInstance = useMapboxInstance()
+  // Used to ensure the style is refreshed on load finish
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  const tilesLoading = useTilesLoading(mapInstance)
+
   if (!mapInstance) return null
 
-  let style: any
+  let style: Style
   try {
     style = mapInstance.getStyle()
   } catch (e) {