+++ applications/amathea/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @globalfishingwatchapp/amathea
 
+## 0.14.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @globalfishingwatch/layer-composer@4.15.0
+  - @globalfishingwatch/react-hooks@4.9.1
+  - @globalfishingwatch/ui-components@2.13.7
+
+## 0.14.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @globalfishingwatch/layer-composer@4.14.5
+  - @globalfishingwatch/react-hooks@4.8.3
+  - @globalfishingwatch/ui-components@2.13.6
+
+## 0.14.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @globalfishingwatch/layer-composer@4.14.4
+  - @globalfishingwatch/react-hooks@4.8.2
+  - @globalfishingwatch/ui-components@2.13.5
+
+## 0.14.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @globalfishingwatch/layer-composer@4.14.3
+  - @globalfishingwatch/react-hooks@4.8.1
+  - @globalfishingwatch/ui-components@2.13.4
+
 ## 0.14.1
 
 ### Patch Changes